requirements.txt
apache-beam==2.46.0
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==1.7.3
fasteners==0.18
gcloud==0.18.3
google-api-core==2.11.0
google-apitools==0.5.31
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-bigquery==3.10.0
google-cloud-bigquery-storage==2.16.2
google-cloud-bigtable==1.7.3
google-cloud-core==2.3.2
google-cloud-datastore==1.15.5
google-cloud-dlp==3.12.1
google-cloud-language==1.3.2
google-cloud-pubsub==2.16.0
google-cloud-pubsublite==1.8.1
google-cloud-recommendations-ai==0.7.1
google-cloud-spanner==3.32.0
google-cloud-storage==2.8.0
google-cloud-videointelligence==1.16.3
google-cloud-vision==3.4.1
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.0
grpcio-status==1.48.2
hdfs==2.7.0
httplib2==0.21.0
idna==3.4
init==0.1.0
numpy==1.24.3
oauth2client==4.1.3
objsize==0.6.1
orjson==3.8.10
overrides==6.5.0
packaging==23.1
pandas==2.0.1
proto-plus==1.22.2
protobuf==3.20.3
pyarrow==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydot==1.4.2
pymongo==3.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
rsa==4.9
six==1.16.0
smart-open==6.3.0
sqlparse==0.4.4
terminal==0.4.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
zstandard==0.21.0
